==> regresion-lineal/07-RL-ML-ANOVA.ipynb <==
import pandas as pd
import statsmodels.api as sm
from statsmodels.formula.api import ols
from statsmodels.multivariate.manova import MANOVA
from sklearn.preprocessing import StandardScaler

df=pd.read_csv('../data/06-primas_norm_est_standard.csv',index_col=0)
df.head(10)

# Pair Programming ANOVA

## Ana C y Ana G

En el pair programming de hoy usaremos el set de datos que guardastéis en el pair programming de normalización y estandarización.

Hasta ahora habéis estado evaluando las características de vuestro set de datos y habéis hecho una gran exploración, es el momento de hacer vuestro primer ANOVA! En el ejercicio de hoy tendréis que hacer un ANOVA con vuestro datos y hacer una interpretación de los resultados.

df.head()

mi_primera_anova = ols('charges_Sklearn ~ age+sex+bmi+children+smoker+region', data=df).fit()
sm.stats.anova_lm(mi_primera_anova)

Si PR es >=0.05, nuetsras variables predictoras influyen en nuestra variable respuesta, por lo que vemos, que todas nuestras variables, influyen 'charges'.

mi_primera_anova.summary()

>- Nos da un R2 de 0.774, es decir, nuestras variables predictoras explican en un 77.4% nuestra variable respuesta.
>- Podemos que ver que todas nuestras variables predictoras excepto sexo (masculino), tienen relación positiva, si aumentan, aumenta el valor de 'charges. 
>- Con p-value<0.05 tenemos edad, bmi, fumadores, rechazamos H0, nuestras variables predictoras tienen efecto en nuestra variable respuesta.
>- La variable de localización no la estamos tomando en cuenta, desde el inicio del pair no nos salen resultados lógicos.
