==> regresion-lineal/08-RL-ML-Encoding.ipynb <==
# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd
import seaborn as sns

# Para la codificación de las variables numéricas
# -----------------------------------------------------------------------
from sklearn.preprocessing import LabelEncoder # para realizar el Label Encoding 
from sklearn.preprocessing import OneHotEncoder  # para realizar el One-Hot Encoding

# Para evitar que salgan los warnings en jupyter
# -----------------------------------------------------------------------
import warnings
warnings.filterwarnings('ignore')

df=pd.read_csv('../data/06-primas_norm_est_standard.csv',index_col=0)
df.head(2)

# Pair Programming Encoding

## Ana C y Ana G

En el pair programming de hoy usaremos el set de datos que guardastéis en el pair programming de normalización y estandarización.

Vuestro set de datos debería tener al menos una variable categórica, el objetivo del pair programming de hoy:

- Hacer una códificación de la/las variables categóricas que tengáis en vuestro set de datos.
- Recordad que lo primero que deberéis hacer es decidir su vuestras variables tienen o no orden, para que en función de esto uséis una aproximación u otra.
- Guardad el dataframe, donde deberíais tener las variables estadandarizas, normalizadas y codificadas en un csv para usarlo en el próximo pairprogramming

df.sample(5)

Tenemos dos, en el caso de hábitos tabáquicos, asumimos que nuestra variable tiene orden, y le asignaremos más peso a la condición fumador.
La variable de sexo también tenemos orden, según ANOVA la condición masculina obtiene un coeficiente negativo y p> a 0.05, nos quedamos con mayor peso a masculino lo haremos por get dummies y seleccionaremos la columna que nos interesa.En el pair 1 vimos que más que el sexo, las primas se relacionaban con el tabaquismo más que por sexo, es decir, eran más los hombres que fumaban y por tanto, se les asocia mayor prima por ello, no por el sexo en si.

mapa = {'no':0, 'yes':1} 

df["fumadores"] = df["smoker"].map(mapa)

df.head()

# lo haremos para la columna "sex"

dummies = pd.get_dummies(df["sex"], prefix_sep = "_", prefix = "sex", dtype = int)
dummies.head(2)

# juntamos los datos condf_dummies

df_dummies = pd.concat([df, dummies], axis = 1)
df_dummies.head(2)

df_dummies['sex_female'].value_counts()

Categorizamos la variable region, que aún no la hemos quitado del dataframe a pesar de que según anova y gráficos no influye aparentemente en nuestra variable respuesta. Asignamos categorias de igual peso con oneHot.

# iniciamos el método de OneHot Encoder
oh = OneHotEncoder()

# hacemos la codificación de los datos para la variable dada 
transformados = oh.fit_transform(df[['region']])

# convertimos nuestro array con la codificación hecha en un dataframe, donde tendremos dos columnas, una para los "Yes" y otra para los "No"
oh_df = pd.DataFrame(transformados.toarray())
oh_df.head(10)

# el método get_feature_names nos va a dar el nombre de las columnas nuevas que se nos generarán
oh_df.columns = oh.get_feature_names_out()# metodo actualizado el anterior lo han quitado ya.
oh_df.columns

# concatenamos el dataframe original con el dataframe que acabamos de crear
final = pd.concat([df_dummies,oh_df],axis=1)
final.head()

final.columns

final.columns=['age','sex','bmi','children','smoker','region', 'charges_Sklearn','fumadores','mujeres','hombres','northeast', 'northwest', 'southeast', 'southwest']
final.head(2)

final.drop('mujeres', axis=1, inplace=True)

final.to_csv('../data/08-primas_categorizadas_normalizadas.csv')
